+++ src/modules/taskClass.js
@@ -1,15 +1,4 @@
-export const taskList = [
-  {
-    description: 'take the trash out',
-    completed: true,
-    index: 0,
-  },
-  {
-    description: 'Do the dishes',
-    completed: false,
-    index: 1,
-  },
-];
+export const taskList = [];
 
 export class Task {
   constructor(description, completed = false, index = taskList.length) {
